refactor(v4): V4RepositoryBase + migrate Calibration (behaviour-preserving)
Extracts the CRUD/soft-delete/LegacyId-deduped-bulk-create logic that was
copy-pasted across every V4 record repository into a shared
V4RepositoryBase<TModel, TEntity>. Concrete repos now supply three static-mapper
bridges (ToEntity/ToDomain/ApplyUpdate) and keep only their type-specific queries.

- New IV4TimeSeriesEntity : IV4Entity exposes the columns the base filters/orders/
  watermarks on (Timestamp, DataSource, Device). EF translates these interface
  members the same way it already does for ITenantScoped.TenantId /
  ISoftDeletable.DeletedAt. Span-shaped types (TempBasal) don't implement it and
  stay off the base.
- Base uses ctx.Set<TEntity>() so no per-type DbSet hook is needed. CountAsync and
  BulkCreateAsync are virtual — the LegacyId-only path lives in the base; the
  SyncId-upsert / DeduplicationService participants will override when migrated.
- CalibrationRepository migrated first (simplest: LegacyId-only, no dedup): 344 -> ~90
  lines, keeping GetByCorrelationId/DeleteBySource/DeleteByTimeRange. Ctor signature
  unchanged (two tests construct it directly).

Verified behaviour-preserving: 27 goldens green vs real Postgres, plus the existing
CalibrationRepositoryBulkCreate (SQLite) and EntryDecomposer unit tests.

WriteOrigin intentionally NOT added yet — adding it after all types are on the base
is one change to the base, not 18 separate repos.
